requirements.txt
anyio==3.6.2
api-display-purposes==0.0.3
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
clarifai==2.6.2
clarifai-grpc==9.1.1
configparser==3.8.1
EasyProcess==1.1
emoji==1.7.0
exceptiongroup==1.1.0
future==0.18.3
googleapis-common-protos==1.58.0
grpcio==1.51.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==2.10
instagrapi==1.16.25
instapy==0.6.16
instapy-chromedriver==2.46
jsonschema==2.6.0
MeaningCloud-python==2.0.0
np==1.0.2
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
plyer==2.1.0
protobuf==3.20.3
pycryptodomex==3.15.0
pydantic==1.9.1
PySocks==1.7.1
python-dotenv==0.21.1
python-telegram-bot==20.1
PyVirtualDisplay==3.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.2
semantic-version==2.10.0
setuptools-rust==1.5.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
webdriver-manager==3.8.5
webdriverdownloader==1.1.0.3
wsproto==1.2.0
